fix: move createBaseColorComponent and createColorEditComponent to litComponentAdapters
